% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-plastics.R
\name{clean_plastics}
\alias{clean_plastics}
\title{Clean the plastics dataset}
\usage{
clean_plastics(dat)
}
\arguments{
\item{dat}{A data frame returned by \code{load_data()}}
}
\value{
A cleaned data frame.
}
\description{
Removes empty or invalid rows, standardizes country names,
and filters out aggregated "Grand Total" rows.
}
\examples{
plastic <- load_data()
clean <- clean_plastics(plastic)
}
