% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R, R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load the Break Free From Plastic dataset}
\usage{
load_data()

load_data()
}
\value{
A tibble with plastic pollution data from TidyTuesday 2021-01-26

A data frame.
}
\description{
Load the Break Free From Plastic dataset

Loads in Plastics Data set
}
\examples{
plastic <- load_data()
}
