% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plastic-lm.R
\name{plastic_lm}
\alias{plastic_lm}
\title{Fit a linear model of plastic totals over time}
\usage{
plastic_lm(dat, country_name = "United States Of America")
}
\arguments{
\item{dat}{A data frame returned by \code{load_data()}}

\item{country_name}{A string specifying the country. Defaults to "United States Of America".}
}
\value{
A named list with the lm model object and a tidy summary tibble.
}
\description{
Fits a simple linear regression of total plastic collected vs. year
for a given country.
}
\examples{
plastic <- load_data()
plastic_lm(plastic, country_name = "China")
}
