% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-plastic-types.R
\name{plot_plastic_types}
\alias{plot_plastic_types}
\title{Plot plastic type breakdown for a country}
\usage{
plot_plastic_types(dat, country_name = "China")
}
\arguments{
\item{dat}{A data frame returned by \code{load_data()}}

\item{country_name}{A string specifying the country. Defaults to "China".}
}
\value{
A ggplot object.
}
\description{
Creates a bar chart showing the count of each plastic type
for a given country.
}
\examples{
plastic <- load_data()
plot_plastic_types(plastic, country_name = "China")
}
