% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-top-polluters.R
\name{plot_top_polluters}
\alias{plot_top_polluters}
\title{Plot the top plastic polluting countries}
\usage{
plot_top_polluters(dat, n = 10)
}
\arguments{
\item{dat}{A data frame returned by \code{load_data()}}

\item{n}{Number of top countries to display. Defaults to 10.}
}
\value{
A ggplot object.
}
\description{
Creates a bar chart of the top N countries by total plastic collected.
}
\examples{
plastic <- load_data()
plot_top_polluters(plastic)
}
