% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-country.R
\name{validate_country}
\alias{validate_country}
\title{Validate that a country exists in the dataset}
\usage{
validate_country(dat, country)
}
\arguments{
\item{dat}{A data frame returned by \code{load_data()}}

\item{country}{A string specifying the country name}
}
\value{
Invisible NULL; throws an error if the country is not found
}
\description{
Validate that a country exists in the dataset
}
