% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country-plastic-prop.R
\name{country_plastic_prop}
\alias{country_plastic_prop}
\title{Summarize plastic type proportions for a country}
\usage{
country_plastic_prop(dat, country = "United States of America")
}
\arguments{
\item{dat}{A data frame returned by \code{load_data()}}

\item{country}{A string specifying the country name. Defaults to "United States".}
}
\value{
A tibble with columns for each plastic type and their proportions
}
\description{
Returns the proportion of each plastic type out of the grand total
for a given country, averaged across all years.
}
